% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_R.R
\name{find_R}
\alias{find_R}
\title{Find R File Contain Query Words}
\usage{
find_R(repo, query, logical = c("all", "any"), local = NULL)
}
\arguments{
\item{repo}{repo}

\item{query}{words}

\item{logical}{all or any}

\item{local}{local dir}
}
\description{
Find R File Contain Query Words
}
\examples{
find_R('yikeshu0611/set','as.numeric')
}
